run: look for platform boot string at beginning of line
When running the same kernel in a VM as on the host system and the kernel
boot message from the VM appears on the log output, the run tool assumes
that the host machine has rebooted unexpectedly. With this commit, an
unexpected reboot is assumed only if the kernel boot message appears at
the beginning of a line.

Fixes #2041